// lib/interfaces.tsx
interface monitor {
  id: string;
  price: number | null;
  screen_size: number | null;
  resolution: string | null;
  refresh_rate: number | null;
  response_time: number | null;
  panel_type: string | null;
  aspect_ratio: string | null;
  createdAt: string | null;
  updatedAt: string | null;
  manufacturer: string | null;
  name: string | null;
}